refactor(LottoService): 비즈니스로직 서비스로 이동
